requirements.txt
altgraph==0.17.4
asgiref==3.8.1
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.4.0
dj-database-url==3.1.0
Django==5.1.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.20.0
Flask==3.1.0
fpdf==1.7.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
pefile==2023.2.7
pillow==12.0.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
pycparser==2.22
pygame==2.6.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PySocks==1.7.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.29.0
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.35.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
